fix: Improved header guard detection
From now on, the header guard syntax idiom allows any text to appear before and after it (it doesn't have to be at the beginning and end of the file). To avoid confusing this idiom with normal preprocessor macros, there must be no valid C ++ code outside of the header guard.